Run AU when "package/*" branch is pushed
Changes in package contents is not verified because AU is triggered manually or
a schedule. So we change to run AU  when "package/*" branch is pushed.

However, side effects like releasing packages to Chocolatey.org or pushing tags
by GitReleases plugin are not desirable, this commit introduces an AU_DRY_RUN
environment variable and enables to toggle side effects.